Harmonise the 5k and 20k extraction with the new hts-create-lm.sh
Right, here's what's what:

HDecode's trigram decoder needs escaped apostrophes, so I'll do that for
everything as Marc did.

HBuild will remove <UNK> using -u '<UNK>'

ngram -unk -limit-vocab ...
Removes <UNK> but does not renormalise.  i.e., it does the same as grep -v, but
changes the ngram counts accordingly.

iss/bin/arpa-man.rb ...
Does basically the same as above, but is hellishly slow.  So that's a non-starter.

LNorm ...
Removes <UNK>, but also renormalises the LM.  By this, I mean it leaves the
ngram probabilities unchanged, but changes the backoff weights.  Milos has also
had trouble with this.

HLMCopy ...
Does the same as ngram/grep.  That's the one to use.

So, I'll do this:

*Replace* my CreateLM5k.sh and Marc's CreateLM20k.sh with a single CreateLMs.sh
that does the same for all language models.  It'll put in escapes, but *not*
try to remove UNK.

*Remove* CreateRecDict.sh as (I confirmed) my dict-man.rb does the same and more.

*Create* iss/bin/hts-create-lm.sh that uses HLMCopy to remove UNK, adds <s> and
</s> to word lists etc.  I'll also *remove* that functionality from decode.sh.
 We'll need to test with the mediaparl stuff to check it doesn't have the same
trouble as LNorm.  There could be an option to just switch it out.

*Create* CreateLM.sh (without the s) that uses hts-create-lm.sh to build
language models for the chosen decoder.